Renamed etcdcache to crcache
